+++ src/qq_lib/core/common.py
@@ -601,24 +601,23 @@ def is_printf_pattern(pattern: str) -> bool:
 
 def split_files_list(string: str | None) -> list[Path]:
     """
-    Split a string containing multiple file paths into a list of Path objects.
+    Split a string containing multiple file paths into a list of relative Path objects.
 
     The string can contain file paths separated by colons (:), commas (,), or
-    any whitespace characters (space, tab, newline). Each path is converted to
-    an absolute Path using Path.resolve().
+    any whitespace characters (space, tab, newline).
 
     Args:
         string (str | None): The string containing file paths. If None or empty,
                              an empty list is returned.
 
     Returns:
-        list[Path]: A list of resolved Path objects corresponding to the individual
-                    file paths in the input string.
+        list[Path]: A list of Path objects corresponding to the individual
+                    relative file paths in the input string.
     """
     if not string:
         return []
 
-    return [Path(f).resolve() for f in re.split(r"[:,\s]+", string)]
+    return [Path(f) for f in re.split(r"[:,\s]+", string)]
 
 
 def to_snake_case(s: str) -> str: