+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Added support for the `-h` flag as a shorthand for `--help` to display the help output.
 - Added autocomplete for qq.
 - Names for loop jobs are now created correctly even if the job script has a file extension.
-- Fixed a bug which caused the archive directory to be created in an incorrect directory when submitting a loop job from a different directory than in which the script was located.
+- Made it possible to submit qq jobs from directories other than the current working directory.
 - `get_info_files_from_job_id_or_dir` now properly catches PermissionError when the user does not have permissions to read an info file.
 - Obtaining the list of jobs on Slurm is now much faster (but can be still slow due to Slurm limitations).
 - Fixed a bug which caused the job to not be able to use multiple MPI ranks on some PBS clusters.